feat: move stock trade form to components
